Move CompositeScale under models/scales
