+++ bokehjs/src/lib/models/scales/index.ts
@@ -1,6 +1,7 @@
 export {CategoricalScale}         from "./categorical_scale"
+export {CompositeScale}           from "./composite_scale"
 export {ContinuousScale}          from "./continuous_scale"
-export {LinearScale}              from "./linear_scale"
 export {LinearInterpolationScale} from "./linear_interpolation_scale"
+export {LinearScale}              from "./linear_scale"
 export {LogScale}                 from "./log_scale"
 export {Scale}                    from "./scale"
